Aligning Space API datatypes

+++ internal/openmetrics-exporter/arrays_space_collector.go
@@ -31,31 +31,31 @@ func (c *ArraySpaceCollector) Collect(ch chan<- prometheus.Metric) {
 		ch <- prometheus.MustNewConstMetric(
 			c.ReductionDesc,
 			prometheus.GaugeValue,
-			as.Space.DataReduction,
+			float64(as.Space.DataReduction),
 			t,
 		)
 		ch <- prometheus.MustNewConstMetric(
 			c.SpaceDesc,
 			prometheus.GaugeValue,
-			as.Space.Snapshots,
+			float64(as.Space.Snapshots),
 			t, "snapshots",
 		)
 		ch <- prometheus.MustNewConstMetric(
 			c.SpaceDesc,
 			prometheus.GaugeValue,
-			as.Space.TotalPhysical,
+			float64(as.Space.TotalPhysical),
 			t, "total_physical",
 		)
 		ch <- prometheus.MustNewConstMetric(
 			c.SpaceDesc,
 			prometheus.GaugeValue,
-			as.Space.Unique,
+			float64(as.Space.Unique),
 			t, "unique",
 		)
 		ch <- prometheus.MustNewConstMetric(
 			c.SpaceDesc,
 			prometheus.GaugeValue,
-			as.Space.Virtual,
+			float64(as.Space.Virtual),
 			t, "virtual",
 		)
 		ch <- prometheus.MustNewConstMetric(
@@ -67,7 +67,7 @@ func (c *ArraySpaceCollector) Collect(ch chan<- prometheus.Metric) {
 		ch <- prometheus.MustNewConstMetric(
 			c.SpaceDesc,
 			prometheus.GaugeValue,
-			as.Capacity,
+			float64(as.Capacity),
 			t, "capacity",
 		)
 
@@ -75,13 +75,13 @@ func (c *ArraySpaceCollector) Collect(ch chan<- prometheus.Metric) {
 			ch <- prometheus.MustNewConstMetric(
 				c.SpaceDesc,
 				prometheus.GaugeValue,
-				as.Capacity-as.Space.TotalPhysical,
+				float64(as.Capacity-as.Space.TotalPhysical),
 				t, "empty",
 			)
 			ch <- prometheus.MustNewConstMetric(
 				c.UtilizationDesc,
 				prometheus.GaugeValue,
-				as.Space.TotalPhysical/as.Capacity*100,
+				float64(as.Space.TotalPhysical)/float64(as.Capacity)*100,
 				t,
 			)
 		}
