+++ internal/rest-client/buckets.go
@@ -8,13 +8,13 @@ type Account struct {
 
 type EradicationConfig struct {
 	ManualEradication string `json:"manual_eradication"`
-	EradicationDelay  int    `json:"eradication_delay"`
+	EradicationDelay  int64  `json:"eradication_delay"`
 }
 
 type ObjectLockConfig struct {
 	Enabled              bool   `json:"enabled"`
 	FreezeLockedObjects  bool   `json:"freeze_locked_objects"`
-	DefaultRetention     int    `json:"default_retention"`
+	DefaultRetention     int64  `json:"default_retention"`
 	DefaultRetentionMode string `json:"default_retention_mode"`
 }
 
@@ -22,14 +22,14 @@ type Bucket struct {
 	Name             string            `json:"name"`
 	Id               string            `json:"id"`
 	Account          Account           `json:"account"`
-	Created          int               `json:"created"`
+	Created          int64             `json:"created"`
 	Destroyed        bool              `json:"destroyed"`
-	TimeRemaining    int               `json:"time_remaining"`
-	ObjectCount      float64           `json:"object_count"`
+	TimeRemaining    int64             `json:"time_remaining"`
+	ObjectCount      int64             `json:"object_count"`
 	Space            Space             `json:"space"`
 	Versioning       string            `json:"versioning"`
 	BucketType       string            `json:"bucket_type"`
-	QuotaLimit       float64           `json:"quota_limit"`
+	QuotaLimit       int               `json:"quota_limit"`
 	HardLimitEnabled bool              `json:"hard_limit_enabled"`
 	RetentionLock    string            `json:"retention_lock"`
 	EradicationCfg   EradicationConfig `json:"eradication_config"`
@@ -38,7 +38,7 @@ type Bucket struct {
 
 type BucketsList struct {
 	CntToken     string   `json:"continuation_token"`
-	TotalItemCnt int      `json:"total_item_count"`
+	TotalItemCnt int32    `json:"total_item_count"`
 	Items        []Bucket `json:"items"`
 	Total        Bucket   `json:"total"`
 }
